Task 002: Fixed ratio
Now the ratio displays as a percentage

+++ src/main/java/acme/forms/AdministratorDashboard.java
@@ -54,7 +54,7 @@ public class AdministratorDashboard {
 	Map<Status,Double>						maximumBudgetOfPatronagesGroupedByStatus;
 	
 	
-	int 									ratioOfArtifactsWithChimpums;
+	double 									ratioOfArtifactsWithChimpums;
 	
 	Map<String, Double> averageBudgetOfChimpumsGroupedByCurrency;
 	Map<String, Double> deviationBudgetOfChimpumsGroupedByCurrency;
